migrate cellcreator to pure & connected components
