+++ packages/core/src/providers/index.js
@@ -1,8 +1,7 @@
 // @flow
-import CellCreator from "./cell-creator";
 import Editor from "./editor";
 import NotebookApp from "./notebook-app";
 import Styles from "./styles";
 import Toolbar from "./toolbar";
 
-export { CellCreator, Editor, NotebookApp, Styles, Toolbar };
+export { Editor, NotebookApp, Styles, Toolbar };
